adding drive mode as a class, refactor assisted navigation iao and drive controller, fix bug with topic /roboy/pinky/middleware/espchair/wheels/dist/front&back
